process_19_12_(2)

+++ src/main/java/org/example/dao_repositories_implements/SubjectDAOImpl.java
@@ -10,37 +10,43 @@ public class SubjectDAOImpl implements SubjectDAO {
 
     DataBase dataBase = DataBase.getDataBase();
 
-    //+!
     @Override
     public ArrayList<Subject> getSubjects() {
         return dataBase.getSubjects();
     }
 
-    //+!
     @Override
     public Subject getSubjectById(int id) {
         return dataBase.getSubjectById(id);
     }
 
-    //+!
     @Override
     public int addSubject(String name) {
         Subject subject = new Subject(0, name);
         return dataBase.addSubject(subject);
     }
 
-    //+!
     @Override
     public String editSubject(int id, String name) {
-        Subject subject = new Subject(id, name);
-        dataBase.editSubject(subject);
-        return "Предмет изменен";
+        Subject subject1 = dataBase.getSubjectById(id);
+        if (subject1 != null) {
+            Subject subject = new Subject(id, name);
+            dataBase.editSubject(subject);
+            return "Предмет изменен";
+        } else {
+            throw new NullPointerException("Такой предмет не найден в системе");
+        }
     }
 
-    //+!
+
     @Override
     public String deleteSubject(int id) {
-        dataBase.deleteSubject(id);
-        return "Предмет удален";
+        Subject subject1 = dataBase.getSubjectById(id);
+        if (subject1 != null) {
+            dataBase.deleteSubject(id);
+            return "Предмет удален";
+        } else {
+            throw new NullPointerException("Такой предмет не найден в системе");
+        }
     }
 }
